Networks now work (in alpha form)
The master boots up and successfully becomes ready with a flannel-run
network.  (as do all nodes labeled mlab/type=cloud)

The platform nodes boot up and successfully become ready with both a
flannel network and a host-local network.  (as do all nodes labeled
mlab/type=cloud)

This is being submitted as a checkpoint.  The index2ip stuff still needs
to be integrated, and multus' ability to connect to etcd to actually use
the CRD needs to be fixed.  However, this code works, with no changes,
to boot up k8s-platform-master in cloud and mlab4.lga0t on the platform
and to get both of them in a state where they can run pods successfully.