ConfigMaps are now write-once. (#271)
* ConfigMaps are now write-once.

When a ConfigMap is updated, a new configmap with a new name is created.
This allows old deployments to use old configurations and new
deployments to use new configurations. It also means that daemonsets and
deployments that use this technique do not have to do the
configmap-reload dance.

Note that this only solves the problem for ConfigMaps in k8s-support that use the new `configmap.jsonnet` library. Other ConfigMaps are on their own, although I encourage people to use this same technique.