Run traceroute-caller v0.8.0 in sandbox and staging
traceroute-caller (TRC) v0.8.0 is functionally the same as TRC
v0.6.0 running on production generating traceroute2 datatype
but with cleaned up source code, paris-traceroute removed,
and latency metrics added.

What was previously running on sandbox and staging was TRC v0.7.x
generating traceroute3 datatype but it had issues because it
was not saving all of scamper output.

Once we've confirmed TRC v0.8.0 works as expected on sandbox
and staging, it will be deployed to production.