Ignores unknown sysctls/changes initContainer name
Not every DaemonSet pod is going to have a "net1" interface, since not
all of them use ipvlan. This commit passes the --ignore flag to sysctl,
causing it to ignore errors about unknown sysctls. It also uses the long
flag for --write, and changes the name of the initContainer to
set-pod-sysctls, which is more intuitive.