Sandbox kinkade (#602)
* Ignores unknown sysctls/changes initContainer name

Not every DaemonSet pod is going to have a "net1" interface, since not
all of them use ipvlan. This commit passes the --ignore flag to sysctl,
causing it to ignore errors about unknown sysctls. It also uses the long
flag for --write, and changes the name of the initContainer to
set-pod-sysctls, which is more intuitive.

* Uses short flags specific to busybox

Busybox does not support long flags, and the short flags it does support
are not the same as the full version on my Linux system.

* Uses a shell to modify ipv6 sysctls

Not knowing up front which interfaces will exist in a pod, statically
configuring the interface to 'net1' will cause the initContainer to fail
in pods where it does not exist. Use a shell instead to take advantage
of shell globbing.