Fix traceroute-caller flags and values when invoking it
Starting with v0.9.0, traceroute-caller supports hop annotations.
The output directory for hop annotations is specified by a new
"hopannotation-output" flag.

The old flag "outputPath" for traceroute results is now called
"traceroute-output" so it will be clear which output it is
referring to.  Also, the value of "traceroute-output" is changed
from "traceroute" to "scamper1" per "Traceroute migration to
Paris1, Scamper1, HopAnnotation1" doc.