Adds new "osupgrade" NDT canary DaemonSet (#627)
* Adds new NDT canary Daemonset 'osupgrade'

This new DaemontSet will run the same NDT version as the regular
production deployment. It is only being leveraged to help filter data so
that we can find out whether upgrading all OS packages, inclulding the
kernel, on the nodes will affect NDT data in any measurable way.

* Removes unsupported -canary flag from osupgrade DS

The flag was added to ndt-server on June 28th, but the current
production release was from June 1, so does not contain this flag yet.