Sets resolvConf field in KubeletConfiguration
A couple times now we have seen a CoreDNS pod crash looping with an
error about a forwarding loop for the root domain. The documentation for
the 'loop' plugin, which detects this, suggests changing the value of
resolvConf in the Kubelet configuration from /etc/resolv.conf to
/run/systems/resolv/resolv.conf on systems that are using
systemd-resolved, which we do.