Fixes incorrect label and adds collection of broadcast metrics (#633)
* Fixes metric name for ifHCInOctets

A "fix" for this was previously done in the disco repository:

https://github.com/m-lab/disco/pull/21

However, I hadn't remembered that that file is just a sample for how to
format the file, and that the real DISCOv2 config file is located in
this repository.

* Adds collection of broadcast packet metrics

The original DISCOv1 used to collect both broadcast and multicast
metrics, but for some reason I chose not to collect those for DISCOv2.
Upon closer inspection, we decided that while multicast packet metrics
were indeed not necessary, that broadcast packet metrics could possibly
be useful. This commit adds collection for ingress and egress broadcast
packets.