Puts "ndtcloud" data into standard GCS bucket
Previously, data for ndtcloud was being pushed to a special bucket in
GCS, a bucket not processed by the ETL. This commit causes pusher to
push ndtcloud data to the standard GCS bucket so that it makes it to the
ETL and eventually BigQuery.

This commit also aims to make the "ndtcloud" DaemonSet more of the form
of our typical canary DaemonSets.