Configures pusher to upload ndt-virtual data to standard GCS bucket (#635)
* Puts "ndtcloud" data into standard GCS bucket

Previously, data for ndtcloud was being pushed to a special bucket in
GCS, a bucket not processed by the ETL. This commit causes pusher to
push ndtcloud data to the standard GCS bucket so that it makes it to the
ETL and eventually BigQuery.

This commit also aims to make the "ndtcloud" DaemonSet more of the form
of our typical canary DaemonSets.

* Sets pcap to listen on eth0 only in host experiment

Previously, any experiment with hostNetwork=true automatically caused
packet-headers to listen on interface eth0. However, there are very few
cases where packet-headers needs to listen explicitly on eth0, other
than the "host" experiment. This commit narrows the conditional to only
cause packet-headers to listen on eth0 for the host experiment.

* Renames "ndtcloud" DaemonSet to "ndt-virtual"

* Renames ndtcloud to ndt-virtual in system.jsonnet