Updates Vector DaemonSet (#638)
* Updates Vector to version 0.19.0-distroless-libc

The previous version of Vector was quite old and had "nighly" in the
version string. We observed a bug in Vector, which upgrading should
hopefully resolve: https://github.com/m-lab/k8s-support/issues/637

* Changes vector version to simplyl 0.19.0

* Uses Vector helm _chart_ version, not app version

* Adds role:Agent to Vector configuration

* Fixes/update to Vector config file

* Changes all = to :, for proper YAML

* Adds condition.type=check_fields to Vector config

* Updates vector transform filter syntax

* Changes Vector config substantially

We had been using a very old version of Vector. The newest version has
quite a lot of changes to the configuration compared to the old one.
This commit reflects the changes I think are necessary to achieve the
desired result.

* Adds fixes to Vector helm deployment

* Adds missing line continuation character

* Uses new syntax for 'remap'-style conditions for Vector

* Uses VRL includes() function in filter condition

* Simplifies kernel_log transform

* Updates comment about K8S_VECTOR_IMAGE

* Specifies Vector's helm chart version