Set -txcontroller.max-rate=150000000 virtual nodes (#646)
This is normally set at runtime on physical machines based the value in
the ConfigMap nodes-max-rate. However, that ConfigMap is created by
looking at the switch configuration for a site, but virtual sites do not
have switches and thus no entry for switches.json in siteinfo. This
commit manually configures the value to the same value we use for other
1g sites.