Adds support for adding IPv6-enabled virtual nodes
Note: dual stack GCE VMs are only available in a few regions. As of this
commit, the only region in the US is us-west2 (LAX). With these changes,
if a VM is being added in one of the few regions where IPv6 is
supported, then this script will automatically enable IPv6.