Adds support for adding IPv6-enabled virtual nodes (#651)
* Adds support for adding IPv6-enabled virtual nodes

Note: dual stack GCE VMs are only available in a few regions. As of this
commit, the only region in the US is us-west2 (LAX). With these changes,
if a VM is being added in one of the few regions where IPv6 is
supported, then this script will automatically enable IPv6.

* Unconditionally writes external-ipv6 metadata file

We want the file to exist for a virtual nodes, even if empty, so that
it is more explicit that a virtual site does or does not support IPv6.