Loads tcp_bbr in a better manner + configure fq qdisc
A previous commit loaded the tcp_bbr module in the script that creates
the VM. This commit moves that logic to the cloud-init file. This commit
also moves coniguring the fq qdisc to the cloud-init file as a systemd
service so that it gets configured on every boot, not just VM creation
time.