Loads BBR module + configures fq qdisc (#653)
* Fixes a small bug in grep for IPv6 support

* Loads BBR module + configures fq qdisc

* Loads tcp_bbr in a better manner + configure fq qdisc

A previous commit loaded the tcp_bbr module in the script that creates
the VM. This commit moves that logic to the cloud-init file. This commit
also moves coniguring the fq qdisc to the cloud-init file as a systemd
service so that it gets configured on every boot, not just VM creation
time.

* Writes NIC iface name to a file for use by tc

We can't be sure that all GCE VMs will have ens4 as the device name, so
these changes write the name of the device to a file and then use that
value to specify which network interface we are configuring.