Removes Vector internal metrics (#655)
* Removes Vector internal metrics

It was decided that we would monitor the functioning of Vector by
looking for logs in Google Logging (Stackdriver), and not by trying to
infer Vector's state by using its own internal metrics. Since nothing
was going to be using these metrics, we just remove them to reduce the
load on Prometheus.