Replaces Vector with fluent-bit (#665)
* First attempt at replacing Vector with fluent-bit

* Fixes a typo in the directory path for fluentbit

* Adds --values flag to helm for fluent-bit install

Also, uses a debug image of fluent-bit, which should have a shell for
debugging.

* Fixes incorrect template substitution for fluent-bit

* Indents 4 spaces and removes empty service: section

* Removes dupicate PARSER config for fluentbit

Also, changes name of secret data for fluentbit

* Splits systemd filters into separate INPUTs

fluent-bit's filtering is not rich enough to support combining && and
|| logic into a single INPUT, so this commit splits the single INPUT
into several others so that we can filter and "kernel" message of
particular priorities.

* Creates [OUTPUT] config in initConatiner

* Creates new fluent-bit-etc volume and mounts

* Fixes typo in volumeMounts field (was missing S)

* Readds outputs field to fluent-bit config with @INCLUDE

* Enables fluent-bit kubernetes filter to add metadata

* Adds double quotes around PRIORITY filter value

* Fixes fluent-bit config indentation

* Allow fluent-bit to access node and node/proxy resources

This is to allow the use of the "Use_Kubelet" feature where fluent-bit
will get k8s metadata for logs from the local kubelet instead of having
to request the data from the API server.

* Removes Use_Kubelet option from fluent-bit

This option currently requires hostNetwork=true, which is a non-starter.

* Adds back CRI parser to fluent-bit config

Earlier I was getting an error that the CRI parser was already defined,
but now I'm seeing an error that the CRI parser is not registered.
fluent-bit docs say that it is built into fluent-bit. We'll see what
happens again.

Also, modifies tags for various inputs and outputs.

* Puts CRI parser into correct config section

* Temporarily simplifies systemd input plugin for debug

* Adds more default options for fluent-bit service

* Adds a stdout output for debugging systemd plugin

* Adds back individual INPUTs for systemd

* Adds missing outputs config section for fluent-bit

* Turns off HTTP_Server + uses prod Docker image

* Tunes memory limits for tail input plugin

Also sets Read_From_Tail for systemd input plugin, which causes the
plugin to read only new entries in journal on startup instead of trying
to ingest the entire journal.

Also, one small unrelated fix to creating virtual sites.

* Adds storage.path to fluent-bit

* Increases Buffer_Size for k8s filter plugin

* Restores the HTTP_Server so k8s live- readiness probes work

* Changes Buf_Max_Size units for consistency

Also adds a comment about why we have so many systemd inputs.