Sets the default flush value to 15s (#670)
The default flush value is 1s. We are beginning to run up against GCP
quotas and limits with regards to requests per minute to the logging
API. Hopefully this change will cause fluent-bit to flush local data to
the Stackdriver output less frequently, causing us to not hit quotas and
limits.