Changes measurementlab.net/type label to just site-type
The label prefix is optional, and, if ommitted, the label is
automatically assumed to be private/custom, so we can safely remove it
and it will be readily apparently what the label means. Change "type" to
"site-type", which is more specific and can safely be a binary of
"physical" or "virtual".