Bumps fluent-bit version to v1.9.3-debug (#678)
v1.9.3 is the latest release, and we use the debug image, which is based
on debian:bullseye-slim base image. Since we are new to using
fluent-bit, when problems arise we really need to be able to get into
the container to debug. The default image is based on "distroless",
which has absolutely no way to interact with the container. There is no
harm in running the debian-slim image for now. When we are confident
that fluent-bit is running flawlessly without any issues, then we can
perhaps revert to the non-debug image.