Adds a POD_NAME env variable to uuid initContainer
Currently, the uuid utility uses the return value of os.Hostname() as a
prefix for the generated UUID. On physical nodes this turns out to the
pod name, but on virtual nodes it is the node's short name (e.g.,
mlab1-abc01). This causes an inconsistency in UUIDs. This commit
introduces a new POD_NAME env variable into each pod, and the uuid
initContainer can use that instead of os.Hostname(). This should make
UUID generation more deterministic and robust and less dependent the
environment, so to speak.

https://github.com/m-lab/k8s-support/issues/687