Sets some limits for the Stackdriver output plugin (#700)
* Sets some limits for the Stackdriver output plugin

We are having some issues with fluent-bit where a flakly or down
Internet connection causes it to stop sending log data, and it never
recovers. This commit introduces a few new configuration settings which
are aimed to workaround this probable bug in fluent-bit, which are also
probably sensible in any case. We don't want fluent-bit to fill up the
filesystem, and in the case of extended outages we don't need to worry
about trying to send data forever. It is okay for us to lose some log
data.