Uses the --enable_metrics cAdvisor flag
We only use a few metrics from cAdvisor. This commit should fairly
dramatically reduce the number of metrics generated by cAdvisor. It also
enables the --whitelisted_container_labels flag to reduce the number of
labels that each metric contains, since we only use a few of the labels
for anything.