Better support for physical/virtual dashboards + updates to cAdvisor config  (#701)
* Updates recording rules to aggregate on site type

This commit adds an additional layer of aggregation on the node label
mlab/type, for "physical" and "virtual". This should allow the
WorkloadOverview dashboard to also switch on site type.

* Adds a new site label to all k8s-pods metrics

* Sets new site label using a better source label

* Removes stray source_labels field

* Removes relabel rule adding 'site' label for DISCO

* Runs kube-state-metrics on the Prometheus server

* Adds a site label to all proxied-workloads targets

* Adds a relabel config to get 'site' for kube-state-metrics

* Uses the --enable_metrics cAdvisor flag

We only use a few metrics from cAdvisor. This commit should fairly
dramatically reduce the number of metrics generated by cAdvisor. It also
enables the --whitelisted_container_labels flag to reduce the number of
labels that each metric contains, since we only use a few of the labels
for anything.

* Uses -disable-metrics instead of -enable-metrics

-enable-metrics was not added to cAdvisor until the v0.40.0, and we are
still on v0.39.3, which is still the latest stable release.

* Uses proper container label names for cAdvisor

* Whitelists 'workload' label for cAdvisor metrics

* Causes fluent-bit to use Google's public DNS servers

* Updates cAdvisor to v0.44.1

Version v0.40.0 introduced the -enable_metrics flag, which is useful for
us, since we only need to enable 3 metrics and leave the rest disabled.
It also introduced finer-grained metrics to allow for more careful
selection of which metrics are enabled or disabled.

* Changes cAdvisor version to v0.44.0

There is currently not a v0.44.1 tagged image.

* Adds a note about a general pattern in recording rules

* Fixes a typo in a comment