Adds new script to cloud-init to configure fq qdisc
We are now wanting to export a metric via node_exporter's textfile
collector that will indicate whether the fq qdisc was properly
configured on a machine. The previous systemd service baked into virtual
machines was very simple and could not create this new metric file. This
commit creates a new script in the node cloud-init configuration which
will configure the fq qdisc and create a metric file. The systemd
service now simply calls this script.