Set Buffer_Size=0 for fluent-bit kubernetes filter
A setting of 0 means there is no limit to the size of the buffer. We
lately ran into a problem where the data returned by the /pods endpoint
of the kubelet was greater than the previous 1MB setting, causing the
buffer to fill and fluent-bit to not be able to properly annotate log
messages. This was the result of hundreds of evicted pods hanging around
due to a separate issue.