Configurations necessary for upgrade to k8s v1.22.15 (#732)
* Applies changes necessary for upgrade to k8s v1.22

* new kubeadm API v1beta3: https://kubernetes.io/docs/reference/config-api/kubeadm-config.v1beta3/
* The feature Dynamic Kubelet Configuration is deprecated and kubelet will report warning when the flag --dynamic-config-dir is used: https://github.com/kubernetes/kubernetes/pull/102966

* Updates k8s to v1.22.15 + upgrades related components

* Increases fluentbit k8s filter buffer size to 2MB

* Set Buffer_Size=0 for fluent-bit kubernetes filter

A setting of 0 means there is no limit to the size of the buffer. We
lately ran into a problem where the data returned by the /pods endpoint
of the kubelet was greater than the previous 1MB setting, causing the
buffer to fill and fluent-bit to not be able to properly annotate log
messages. This was the result of hundreds of evicted pods hanging around
due to a separate issue.

* Creates k8s manifests dir to quiet down kubelet