Removes maxrate setting for fq qdisc on virtual nodes (#735)
* Renames add "virtual_site" script to "virtual_node"

The script more accurately adds one virtual node to a logical "site".
"virtual_site" made sense when we only had a single node per virtual
site, but now we can have many nodes.

Also, updates the USAGE var in the script to reflect a recent change
where an operator can pass in a machine name.

* Sets MACHINE_TYPE="" if not passed

This avoids an unbound variable error.

* Removes maxrate setting from config for fq qdisc

It was decided that on virtual nodes setting maxrate was not necessary
and the default is better.

* Uses variable instead of static net device name