Uses max-rate value created by max-rate.service
Some time back we introduced a new systemd service named
max-rate.service, which, at boot, downloads a siteinfo format named
max-rate, and populates the file /var/local/metadata/iface-max-rate,
which reflects the appropriate setting for the -txcontroller.max-rate
flag of the ndt-server on a particular node. This commit uses the value
in that file to configure the ndt-server, retiring all the wonky
machinery to make that same thing work from a ConfigMap.