Uses standard metadata object for all ndt manifest
Previously, the ndt-virtual DaemonSet was using a local variable named
"metadata", which contained a Volume and VolumeMount for the host
directory /var/local/metadata, where a number of metadata files exist
which are needed for the ndt-virtual DaemonSet. Now that we are going to
start using the file /var/local/metadata/iface-max-rate to configure the
ndt-server -txcontroller.max-rate flag, it makes sense to move that
"metadata" object to the template file, and to reference that in every
ndt manifest.