Makes use of max-rate value on local machine (#737)
* Uses max-rate value created by max-rate.service

Some time back we introduced a new systemd service named
max-rate.service, which, at boot, downloads a siteinfo format named
max-rate, and populates the file /var/local/metadata/iface-max-rate,
which reflects the appropriate setting for the -txcontroller.max-rate
flag of the ndt-server on a particular node. This commit uses the value
in that file to configure the ndt-server, retiring all the wonky
machinery to make that same thing work from a ConfigMap.

* Adds missing commas in ndt jsonnet files

* Uses standard metadata object for all ndt manifest

Previously, the ndt-virtual DaemonSet was using a local variable named
"metadata", which contained a Volume and VolumeMount for the host
directory /var/local/metadata, where a number of metadata files exist
which are needed for the ndt-virtual DaemonSet. Now that we are going to
start using the file /var/local/metadata/iface-max-rate to configure the
ndt-server -txcontroller.max-rate flag, it makes sense to move that
"metadata" object to the template file, and to reference that in every
ndt manifest.

* Fixes capitalization of Metadata variable

* Fixes capitalization of Metadata var in ndt-virtual

* Adds missing 'exp.' prefix to Metadata var