Limits pcap container to 1500M of memory in cloud
We have had some issues with packet-headers consuming too much memory
and crashing virtual nodes. This appears to happen as a result of SYN
floods. Currently our virtual machines are of type n{1,2}-highcpu-4, and
have only 4G of RAM. This will restrict packets-headers to roughly 1/3
of the available memory, which _should_ be plenty except when something
wrong is happening.