Adds siteType parameter to ExperimentNoIndex()
For now, we need to limit how much memory pcap can use in virtual nodes
due to this: https://github.com/m-lab/ops-tracker/issues/1682

This commit adds resource limits to the pcap container, but only on
virtual nodes. To plumb this it was, unfortunately, necessary to add a
new "siteType" parameter to the ExperimentNoIndex() function.