Adds memory limit of 1500M to pcap on virtual nodes (#741)
* Limits pcap container to 1500M of memory in cloud

We have had some issues with packet-headers consuming too much memory
and crashing virtual nodes. This appears to happen as a result of SYN
floods. Currently our virtual machines are of type n{1,2}-highcpu-4, and
have only 4G of RAM. This will restrict packets-headers to roughly 1/3
of the available memory, which _should_ be plenty except when something
wrong is happening.

* Adds siteType parameter to ExperimentNoIndex()

For now, we need to limit how much memory pcap can use in virtual nodes
due to this: https://github.com/m-lab/ops-tracker/issues/1682

This commit adds resource limits to the pcap container, but only on
virtual nodes. To plumb this it was, unfortunately, necessary to add a
new "siteType" parameter to the ExperimentNoIndex() function.

* Uses singles quotes around variable value

* Increases memory limit for pcap to 2G for ndt-virtual