Updates fluent-bit to v2.0.6 (#749)
Also, take advantage of some new "threaded" feature, whereby input
plugins can run in a seprate thread and not back up the main thread.
Also, reduces storage.total_limit_size to 250M. Our usage of container
logs in Stackdriver is more of a nice-to-have than mission-critical and
we cannot lose even 1byte of log data. We don't need to buffer 1G of
logs if there are network issues, and 250M should be adequate.