Replaces fluent-bit with Vector (#754)
* Replaces fluent bit with Vector

This is more or less the inverse of this close PR from about 9 months
ago: https://github.com/m-lab/k8s-support/pull/665

fluent-bit has _not_ turned out to be more reliable than Vector, or at
least not any better. fluent-bit's reliability today is at least as bad
Vector's lack of reliability was a year ago, and Vector's configuration
and documentation is worlds better than fluent-bit. The bugs that seemed
like they might be responsible for the unreliability of Vector have all
apparently been resolved. Let's try Vector, again.

* Fixes a syntax error and removes condition.type

The default type is "vrl", and a shorthand is to just make the value of
condition be a string that represents the filter.

* Adds a data_dir config for Vector

* Fixes the VectorMissing alert

Vector is not configured to export any metrics, so the usual
`up{deployment="vector"}` expression does not work. However,
kube-state-metrics still knows about it.

* Moves helm values override files to top level dir

Before the helm values override files were mixed in with various jsonnet
config files in the config/ directory. Since helm is its own thing, this
commit creates a new top-level directory named "helm/" and moves all the
value override files to seprate subdirectories of that directory.