Removes privileged=true for flannel containers
The ROS security audit of the M-Lab platform noted that flannel
containers were running with CAP_SYS_ADMIN, and that we should determine
if we could remove that capability. This was the result of setting
privileged=true in the securityContext for the pods. Looking at the
current documentation for flannel, their sample configurations explicity
set privileged=false, and grant only NET_ADMIN and NET_RAW capabilities.
This commit follows this same pattern for M-Lab's flannel pods.