Explicitly sets JSON credential file path
Previously we were setting the env variable
GOOGLE_APPLICATION_CREDENTIALS. The documentation for this sink says
that in the absense of this configuration it will check that env
variable. And indeed it appears to work that way, but over the weekend
around 8 Vector pods started throwing errors about invalid tokens, when
we don't even use tokens to authenticate Vector. This made me feel like
there could be some bug in Vector by which it was falling back to trying
to authenticate using a token or key it tried to fetch from a GCP
metadata server, or something along those lines. This commit explicitly
sets the path to the JSON key for authentication.