Allows cert-manager startupapicheck to run anywhere
Before, it could not run on a control plane node due to taints. This
just allows it to run anywhere, which is useful when a cluster is
initialzed and there are not other nodes than control plane nodes.

Also, use the "installCRDs" option of the helm config for cert-manager
and removed manually installing them in the apply_k8s_configs.sh script.