Adds kubectl apply of system.json earlier
For some reason, when initializing the cluster directly on control plane
nodes using the Packer images, cert-manager startupapicheck was failing
and causing the script to time out. It has something to do with flannel
networking missing on a brand new cluster. This commit moves the
application of system.json to the cluster earlier in the script (which
deploys flannel), before cert-manager is installed by helm.