Don't evaluate setup_k8s.sh template
Evaluation of this template will now happen directly on the first
control plane node when it initializes the cluster.  The template file
will now reside in epoxy-images.

Also, don't try to copy the admin.conf kubeconfig from GCS, but instead
try to login to one of the control plane nodes to get it.