Adds back build step of copying admin.conf from GCS
It was removed, thinking that Cloud Build might be able to login to a
control plane node and get the admin credentials, but that failed. For
now, the cluster init process will continue to upload admin.conf to GCS,
and it can be downloaded by Cloud Build to deploy cluster workloads.
However, this should be fixed so that we don't have to upload admin
cluster crendentials to GCS.