Remove ./node/ subdirectory
The main thing contained in ./node/ was the script template
setup_k8s.sh.template, which was evaluated by the create_k8s_configs.sh
script, and then uploaded to GCS. However, that process now happens in
the custom Packer image for control plane nodes when the control plane
is initialized. These files are no longer needed in this repo.