Runs most containers as non-root (#785)
* First attempt at running most ndt containers as non-root

packet-headers will continue to run as root, but with as few
capabilities as possible.

* Adds missing closing square bracket to array

* Wraps VolumeMount() in volumeMount[]

* Adds missing comma to templates.jsonnet

* Moves function _after_ VolumeMount() in file

* Corrects call to setDataDirOwnership()

* Removes quotes from runAs securityContext values

* Nests securityContext correctly in pod spec

* Fixes name of CAP for packet-headers pod

* Removes conditional around setDataDirOwnership()

* Fixes typo in function name

* Makes Wehe mountpath the standard one for M-lab

To make this change possible, this PR had to be merge upstream:

https://github.com/NEU-SNS/wehe-py3/pull/8

* Reenables envelope access tokens for Wehe

* Runs access as root in Wehe

The access envelope needs to modify iptables rules. Run the process as
root, and restrict the capabilities to only NET_ADMIN.

* Runs ndt-virtual and responsiveness as root

Both of these pods run with hostNetwork=true, and the kubelet will
refuse to set net.ipv4.ip_unprivileged_port_start=80 when this is true.
Instead, run these pods as root, and remove all capabilities except for
CAP_NET_BIND_SERVICE.

* Excludes ndt-virtual and responsiveness from sysctl

ndt-virtual and responsiveness run with hostNetwork=true, and in that
case k8s will not apply the unprivileged_port_start sysctl. For those,
the containers run as root with only the capability
cap_net_bind_service.

* If hostNetwork=true, then don't set port sysctl

When hostNetwork=true k8s will refuse to set the unprivileged port start
sysctl. In those cases, run the containers as root with only the
cap_net_bind_service capability.

* Runs disco as a non-root user

* Fixes jsonnet syntax error, replace comma with semicolon

* Adds missing comma in Disco jsonnet file

* Makes disco initContainer an array, not map

* Runs node-exporter as non-root user

* Makes disco use setDataDirOwnership() function

Don't replicate the setDataDirOwnership() return value in the disco
manifest, but instead just call the function.

* Changes ownership of node-exporter data dir

* Runs cAdvisor as non-root, removes 'utilization' DS

The "utilization" DaemonSet was defunct and no longer used, but for some
reason the manifest was still hanging around.

* Changes autoloading datatypes dir ownership

This directory is different from where experiments store data
(/cache/data/<experiment>), and is today /var/spool/datatypes. In any
case, nodeinfo (currently the only autoloading experiment) runs as user
nobody and needs to be able to write to this directory.

* Changes api- back to master- for control plane node

* Runs cAdvisor as root, but with limite capabilities

I tried running it as non-root, but it didn't work. The containers
didn't crash, but many metrics disappeared. Presumably there are files
and directories that cAdvisor needs to read which only root can do. Run
as root, but with only a single capability: CAP_DAC_READ_SEARCH.

* Moves securityContext{} to container level, not pod

* Adds missing curly brace to cadvisor manifest

* Runs Pusher as root, but with limited capabilities

Some containers (e.g., packet-headers) run as root, and write data files
owned by root. Pusher needs to be able to work with these files, so run
Pusher as root, but with only the CAP_DAC_OVERRIDE capability.