Run most containers as non-root (#795)
* Uses user and group numbers instead of names

Not every container images filesystem uses the same user and group name
to number mappings. Use numbers instead of names to avoid permission
issues.

Additionally, when a container needs to run as root, don't also run as
group 0, but instead use the pod default. This allows us to leverage
group file permissions on data directories.

* Creates datatype dirs in initContainer

* Adds datatypes param to exported function

* Passes datatypes param to setDataDirOwnership for Disco

* Uses -p flag to mkdir to avoid errors

Also, updates call to setDataDirOwnership for node-exporter to pass in
an empty array for datatypes, since node-exporter doesn't have
datatypes.

* Passes dot to setDataDirOwnership() for node-exporter

node-exporter has not datatypes. Passing dot will result in "mkdir -p
.", which is a no-op.

Adds to comment on setDataDirOwnership() about why we are creating
datatype directories in advance.

* Runs traceroute-caller as root

Apparently scamper want to chroot to /var/empty, which is a privileged
operation. Run traceroute-caller as root, but with only the
CAP_SYS_CHROOT capability.

* Adds missing comma to jsonnet file

* Adds CAP_DAC_OVERRIDE to traceroute-caller

It's needs CAP_SYS_CHROOT because scamper apparently wants to chroot to
/var/empty now it needs CAP_DAC_OVERRIDE to be able to connect to the
tcp-info event socket, which is owned by nobody:nogroup. I'm not sure
why CAP_DAC_READ_SEARCH is not sufficient for this, but it isn't.

* Updates Pusher to v1.20.3

The only change in this new version is that it will not try to create
the datatype data directories if they do not already exist. Instead, let
each service that creates the datatype create the directory.

* Gives traceroute-caller setuid and setgid capabilities

* Removes datatypes param from setDataDirOwnership()

All datatypes now create the directories they needs, and Pusher no
longer creates datatype directories, so permissions and ownership show
be handled correctly now.

* Gives traceroute-caller all default capabilities

Scamper requires too many capabilities to make it worth trying to tease
out exactly which ones it needs and doesn't.

* Imports templates.jsonnet into node-exporter manifest

* Adds pod securityContext for user:group 65534

* Revert "Revert "Runs most containers as non-root (#785)" (#788)"

This reverts commit 4f1fbabc4ab7e945a1f42b0144596b100bf51475.

* Adds tmp dir env var for matplotlib to wehe

The default directory is only writable by root, and matplotlib was
logging warnings about how not setting this env dir would slow things
down.

* Runs wehe as root with NET_RAW capability

Wehe needs to do packet captures, which apparently require the NET_RAW
capability, just like the packet-headers container.

Also, better sort manifest fields alphabetically.

* Bumps wehe version to v0.2.8

* Updates traceroute-caller and packet-headers versions

The new versions have binaries with necessary file capabilities applied
so that the container processes can run as non-root.

* Updates wehe to v0.2.9

This new version allows wehe to run as a non-root user.

* Runs Pusher as root

The ndt-virtual and responsiveness experiments still run as root
(because they are hostNetwork=true and need to bind to port 80). Pusher
needs to run as root to work with data from these experiments as well as
data owned by nobody.

* Adds missing common to jsonnet file

* Explicitly adds needed capabilities and drops others

It would appear that a container process running as non-root will get no
capabilities by default. However, the container maintains a set of
potential capabilities that processes could have. This commit explicitly
drops all capabilities for a container, and explicitly allows only ones
possibly needed by the container.

* Fixes typo in capabilities in several files

* Starts check-reboot.timer on VM creation

* Moves securityContext to right place in ndt manifest

* Adds securityContext to Vector values-overrides

Run Vector as root, but with only the DAC_READ_SEARCH capability.

* Updates revtr to container image version v0.3.1

* Adds necessary capabilities to revtr container

All processes run as nobody:nogroup, but the scamper binary needs these
capabilities.

* Removes defunct "utilization" experiment from expressions

* Explicitly drops all capabilities for kube-rbac-proxy

* Changes square brackets to curly braces

* Adds TODO comment to responsiveness manifest

* Removes -R from chmod

Most of the files under the data dir will be regular files, for which
those modes would not be suitable.

* Fixes a typo in a comment

* Removes unnecessary text from comment

* Removes MPLCONFIGDIR env variable from wehe

This is a configuration that should be baked into the container image
from the upstream repo, not added to our k8s configs.

* Removes responsiveness experiment, for now

Since it's not a first class experiment yet it was pinned to sandbox,
and had hostNetwork=true, since it didn't have an experiment slot with
a public IP address. And with hostNetwork=true, the container had to run
as root so that it could bind to port 80.

* Adds comment to revtr about removing CHOWN later

* Uncomments startupProbe for Wehe

* Removes duplicate securityContext block for wehe