Don't run wehe containers as root
I thought that this should have already been the case, but was surprised
to see wehe processes running as root in staging, and then getting
permission denied errors to write to /var/spool/wehe/replay (owned by
nobody:nogroup). I had to manually resolve numerous conflicts manually
in the wehe manifest when I rebased my "non-root" branch changes on
main, and it's not impossible that I made a mistake in there.