Fix ndt-virtual's permission issue (#804)
* Set data dir permissions to 2755 recursively.

This means that the setgid bit is applied to datatype folders that already exist
when the initContainer is started, such as those mounted separately from the
main /var/spool/<experiment> folder.

For example, /var/spool/ndt/ndt7 is mounted as a separate volume in the
ndt-server container. Kubernetes will create this folder as owned by root/root
and with 0755 permissions. Then, the set-data-dir-perms initContainer sets the
owner to nobody/nogroup. This would not allow any other user to write to this
folder.

* Revert "Do not drop privileges in ndt-virtual's ndt containers (#802)"

This reverts commit 079be284b254d1cb4fb3240207bd0e0b5ad6f4f7.

* Run as group 'nogroup'.

* Do not chmod recursively, the top-level folders are enough.

* Do not overwrite pod-level runAsGroup