Fixes bug in -token.machine flag value (#825)
Prior to this commit, the shell parameter substitution creating the
value for the -token.machine flag would strip off "-lab.org" from VMs
that were not part of a MIG. The original thinking had been that we
would only be using MIGs for all VMs, but we decided that we have use
cases for both MIGs and regular stand-alone VMs. This commit uses a
parameter subtitution that should work in the case of a MIG instances,
and be a no-op on non-MIG instances.