Use ndt-virtual manifest as template for ndt-canary
The previous ndt-canary manifest assumed it was being deployed on a
physical machine. This canary is now designed to be deployed on an
Equinix machine. Rather than try to find all the places where the
existing canary manifest differed from the virtual one, I just copied
the ndt-virtual manifest onto the ndt-canary one, and then made edits as
necessary.