Updates Kured and Helm chart versions
Within the last day, Kured released a new chart which defaults to
hostNetwork=false. It is also includes a configuration allowing the user
to select the metrics port to listen on, which defaults to 8080. This
commit sets the metriscPort to 9092 to be more consistent with ports we
use for other metrics endpoints.