Disable Vector kubernetes_logs for pods by default
After further discussion we realized that there is probably no reason to
be pushing any container logs to GCP, by default. Today, the are
essentially unused. This commit changes the remaining experiments pods
to have the label "vector.dev/exclude: true". In the event that we need
container logs persisted for longer than they are in the cluster, then
we can easily just modify this label to enable Vector to push logs to
GCP.